requirements.txt
tensorflow-macos
torch==2.0.1
torchvision==0.15.2
scikit-learn==1.2.2
pandas==2.0.1
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
opencv-python==4.7.0.72
Flask==2.3.2
Pillow==9.5.0
tqdm==4.65.0
albumentations==1.3.0
gradio==3.32.0